#56: refactor video portal module base package
