+++ webcentral-backend-videoportal/src/main/java/hu/bme/sch/bss/webcentral/videoportal/domain/VideoResponse.java
@@ -1,7 +1,7 @@
-package hu.bme.sch.bss.webcentral.domain;
+package hu.bme.sch.bss.webcentral.videoportal.domain;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import hu.bme.sch.bss.webcentral.model.Video;
+import hu.bme.sch.bss.webcentral.videoportal.model.Video;
 
 @JsonSerialize
 public final class VideoResponse {
