+++ webcentral-backend-videoportal/src/test/java/hu/bme/sch/bss/webcentral/videoportal/domain/VideoResponseTest.java
@@ -1,6 +1,6 @@
-package hu.bme.sch.bss.webcentral.domain;
+package hu.bme.sch.bss.webcentral.videoportal.domain;
 
-import hu.bme.sch.bss.webcentral.model.Video;
+import hu.bme.sch.bss.webcentral.videoportal.model.Video;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
